Added src/net

+++ src/common/substitute.cc
@@ -122,7 +122,7 @@ void SubstituteAndAppend(std::string* output, const char* format,
     }
   }
 
-  DCHECK_EQ(target - output->data(), output->size());
+  DCHECK_EQ(static_cast<size_t>(target - output->data()), output->size());
 }
 
 }  // namespace ncode
